requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.3
blis==0.7.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cmudict==1.0.32
colorama==0.4.6
comtypes==1.4.10
confection==0.1.5
cymem==2.0.11
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
filetype==1.2.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-anthropic==0.3.10
langchain-community==0.3.16
langchain-core==0.3.51
langchain-google-genai==2.1.2
langchain-ollama==0.3.1
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langchain-together==0.3.0
langcodes==3.5.0
langsmith==0.3.25
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.3.2
murmurhash==1.0.12
mypy-extensions==1.0.0
ollama==0.4.7
openai==1.71.0
orjson==3.10.16
packaging==24.2
preshed==3.0.9
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.9.2
pydantic-settings==2.8.1
pydantic_core==2.23.4
Pygments==2.19.1
pyphen==0.17.2
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQtChart==5.15.7
PyQtChart-Qt5==5.15.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyttsx3==2.90
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
structlog==25.2.0
tenacity==9.1.2
textstat==0.7.5
thinc==8.2.5
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.13.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
Wikipedia-API==0.8.1
wrapt==1.17.2
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
selenium
webdriver-manager
